어플리케이션 스탑 테스트내용 수정

+++ scripts/application_stop.sh
@@ -4,12 +4,9 @@ LOG_DIR=/home/ubuntu/build
 LOG_FILE=$LOG_DIR/stop.log
 
 # Ensure the log directory exists
-if [ ! -d "$LOG_DIR" ]; then
-  mkdir -p $LOG_DIR
-  mkdir -p $LOG_DIR/scripts
-fi
+mkdir -p $LOG_DIR
 
-echo "Stopping application at $(date)" >> $LOG_FILE
+echo "Stopping application at $(date '+%Y-%m-%d %H:%M:%S')" >> $LOG_FILE
 
 CURRENT_PID=$(pgrep -f 'sportsLink-.*-SNAPSHOT.jar')
 
@@ -17,8 +14,13 @@ if [ -z "$CURRENT_PID" ]; then
   echo "No application is currently running." >> $LOG_FILE
 else
   echo "Stopping application with PID: $CURRENT_PID" >> $LOG_FILE
-  sudo kill -9 $CURRENT_PID
+  sudo kill -15 $CURRENT_PID
   sleep 5
+
+  if kill -0 $CURRENT_PID > /dev/null 2>&1; then
+    echo "Application did not terminate gracefully, forcing shutdown." >> $LOG_FILE
+    sudo kill -9 $CURRENT_PID
+  fi
 fi
 
-echo "Application stopped at $(date)" >> $LOG_FILE
+echo "Application stopped at $(date '+%Y-%m-%d %H:%M:%S')" >> $LOG_FILE
